# src/api/rest/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{id}": {
      "get": {
        "operationId": "getProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project was not found."
          }
        },
        "summary": "Get a project",
        "tags": ["Projects"]
      },
      "delete": {
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "Project was not found."
          }
        },
        "summary": "Delete a project",
        "tags": ["Projects"]
      }
    },
...
    "/api/projects/{projectId}/requirements": {
      "post": {
        "operationId": "createRequirement",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project was not found."
          }
        },
        "summary": "Create a requirement",
        "tags": ["Projects"]
      },
      "get": {
        "operationId": "listRequirements",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "first",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project was not found."
          }
        },
        "summary": "List requirements",
        "tags": ["Projects"]
      }
    },
    "/api/projects/{projectId}/requirements/{requirementId}": {
      "get": {
        "operationId": "getRequirement",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requirementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record was not found."
          }
        },
        "summary": "Get a requirement",
        "tags": ["Projects"]
      },
      "put": {
        "operationId": "updateRequirement",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requirementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record was not found."
          }
        },
        "summary": "Replace a requirement",
        "tags": ["Projects"]
      },
      "delete": {
        "operationId": "deleteRequirement",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requirementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "Record was not found."
          }
        },
        "summary": "Delete a requirement",
        "tags": ["Projects"]
      }
    },
    "/api/projects/{projectId}/work-plans": {
      "post": {
        "operationId": "createWorkPlan",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkPlanRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project was not found."
          }
        },
        "summary": "Create a work plan",
        "tags": ["Projects"]
      },
      "get": {
        "operationId": "listWorkPlans",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "first",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPlanConnectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project was not found."
          }
        },
        "summary": "List work plans",
        "tags": ["Projects"]
      }
    },
    "/api/projects/{projectId}/work-plans/{workPlanId}": {
      "get": {
        "operationId": "getWorkPlan",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workPlanId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record was not found."
          }
        },
        "summary": "Get a work plan",
        "tags": ["Projects"]
      },
      "put": {
        "operationId": "updateWorkPlan",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workPlanId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkPlanUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Record was not found."
          }
        },
        "summary": "Replace a work plan",
        "tags": ["Projects"]
      },
      "delete": {
        "operationId": "deleteWorkPlan",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workPlanId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "Record was not found."
          }
        },
        "summary": "Delete a work plan",
        "tags": ["Projects"]
      }
    },
...
  },
  "info": {
    "title": "Revo Core API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ProjectResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["id", "name"]
      },
...
      "PageInfoResponse": {
        "type": "object",
        "properties": {
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "startCursor": {
            "type": "string"
          },
          "endCursor": {
            "type": "string"
          }
        },
        "required": ["hasNextPage", "hasPreviousPage"]
      },
...
      "RequirementRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["proposed", "accepted", "deferred", "rejected"]
          },
          "statement": {
            "type": "string"
          },
          "acceptance": {
            "type": "string"
          },
          "relatedAdr": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["id", "title", "status", "statement", "acceptance"]
      },
      "RequirementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["proposed", "accepted", "deferred", "rejected"]
          },
          "statement": {
            "type": "string"
          },
          "acceptance": {
            "type": "string"
          },
          "relatedAdr": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["id", "title", "status", "statement", "acceptance", "relatedAdr"]
      },
      "RequirementEdgeResponse": {
        "type": "object",
        "properties": {
          "cursor": {
            "type": "string"
          },
          "node": {
            "$ref": "#/components/schemas/RequirementResponse"
          }
        },
        "required": ["cursor", "node"]
      },
      "RequirementConnectionResponse": {
        "type": "object",
        "properties": {
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RequirementEdgeResponse"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfoResponse"
          },
          "totalCount": {
            "type": "number"
          }
        },
        "required": ["edges", "pageInfo", "totalCount"]
      },
      "RequirementUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["proposed", "accepted", "deferred", "rejected"]
          },
          "statement": {
            "type": "string"
          },
          "acceptance": {
            "type": "string"
          },
          "relatedAdr": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["title", "status", "statement", "acceptance"]
      },
      "WorkPlanRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "ready", "closed"]
          },
          "outcome": {
            "type": "string"
          },
          "bounds": {
            "type": "string"
          },
          "baselineId": {
            "type": "string"
          },
          "acceptance": {
            "type": "string"
          }
        },
        "required": ["id", "title", "status", "outcome", "bounds", "acceptance"]
      },
      "WorkPlanResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "ready", "closed"]
          },
          "outcome": {
            "type": "string"
          },
          "bounds": {
            "type": "string"
          },
          "baselineId": {
            "type": "string"
          },
          "acceptance": {
            "type": "string"
          }
        },
        "required": ["id", "title", "status", "outcome", "bounds", "baselineId", "acceptance"]
      },
      "WorkPlanEdgeResponse": {
        "type": "object",
        "properties": {
          "cursor": {
            "type": "string"
          },
          "node": {
            "$ref": "#/components/schemas/WorkPlanResponse"
          }
        },
        "required": ["cursor", "node"]
      },
      "WorkPlanConnectionResponse": {
        "type": "object",
        "properties": {
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkPlanEdgeResponse"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfoResponse"
          },
          "totalCount": {
            "type": "number"
          }
        },
        "required": ["edges", "pageInfo", "totalCount"]
      },
      "WorkPlanUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "ready", "closed"]
          },
          "outcome": {
            "type": "string"
          },
          "bounds": {
            "type": "string"
          },
          "baselineId": {
            "type": "string"
          },
          "acceptance": {
            "type": "string"
          }
        },
        "required": ["title", "status", "outcome", "bounds", "acceptance"]
      },
...
    }
  }
}